+++ src/adapter/spi/drivers/FetcherSpi.ts
@@ -3,6 +3,7 @@ import type { IFetcherSpi } from '/domain/services/Fetcher'
 export interface IFetcherDriver {
   get: (url: string, options?: RequestInit) => Promise<Response>
   post: (url: string, body?: object, options?: RequestInit) => Promise<Response>
+  put: (url: string, body?: object, options?: RequestInit) => Promise<Response>
 }
 
 export class FetcherSpi implements IFetcherSpi {
@@ -14,5 +15,9 @@ export class FetcherSpi implements IFetcherSpi {
 
   post = (url: string, body?: object, options?: RequestInit) => {
     return this._driver.post(url, body, options)
+  }
+
+  put = (url: string, body?: object, options?: RequestInit) => {
+    return this._driver.put(url, body, options)
   }
 }