+++ test/automations/actions/code/runTypescript/packages.test.ts
@@ -446,6 +446,49 @@ mock.request(({ app, request }) => {
       expect(response.exist).toBeTruthy()
     })
 
+    it('should run a Typescript code with sodium package', async () => {
+      // GIVEN
+      const config: Config = {
+        name: 'App',
+        version: '1.0.0',
+        automations: [
+          {
+            name: 'sodium',
+            trigger: {
+              service: 'Http',
+              event: 'ApiCalled',
+              path: 'sodium',
+              output: {
+                exist: {
+                  boolean: '{{runJavascriptCode.exist}}',
+                },
+              },
+            },
+            actions: [
+              {
+                service: 'Code',
+                action: 'RunTypescript',
+                name: 'runJavascriptCode',
+                code: String(async function (context: CodeRunnerContext) {
+                  const {
+                    packages: { sodium },
+                  } = context
+                  return { exist: !!sodium.ready }
+                }),
+              },
+            ],
+          },
+        ],
+      }
+      const { url } = await app.start(config)
+
+      // WHEN
+      const response = await request.post(`${url}/api/automation/sodium`)
+
+      // THEN
+      expect(response.exist).toBeTruthy()
+    })
+
     it('should run a Typescript code with Notion package', async () => {
       // GIVEN
       const config: Config = {